[FIX - FEATURE] Grammar compatible with LALR algorithm and interactive mode
File standard_syntax.py

Some rules in the grammar have been modified to make the grammar compatible with
LALR algorithm. The collisions are removed. There is still a bug to fix.

parser() function :
Added the "interactive" parameter and adapted the function definition to use the
auto-completion functionality.

parser_interactive() function :
Performes the auto-completion.